chore: address issue #157
